a more structured approach of the static lookup map
so far the transport map was a pure postfix implementation and also
it was rather clumsy structured.
Having a yaml file with the onion address as key and the email domains
mapped to that onion address as values, makes it more structured,
easier to update and also suitable as an input for maps of MTAs
other than postfix.