changed the address of onion service for my host
this source file is not completely up to date -- we still have a couple of onion service v2 addresses, but there might still be folks using it instead of auto-discovery (I've only just implemented auto-discovery this month on my personal server). I'll keep the address for my server up to date in case some ppl are still generating their transport maps out of it.